pyproject.toml
[project]
name = "learning-pocketflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.58.2",
    "pocketflow>=0.0.2",
    "python-dotenv>=1.1.1",
]
